+++ commerce/payment/src/main/resources/schema.sql
@@ -2,8 +2,8 @@ CREATE TABLE IF NOT EXISTS payments
 (
     payment_id     UUID PRIMARY KEY,
     order_id       UUID    NOT NULL,
-    total_payment  REAL,
-    delivery_total REAL,
-    fee_total      REAL,
+    total_payment  DECIMAL,
+    delivery_total DECIMAL,
+    fee_total      DECIMAL,
     payment_state  VARCHAR NOT NULL
 );
